add support for result checks
