--- src/main/scala/cz.cvut.fit/gatling/protocol/GremlinComponents.scala
@@ -1,9 +1,0 @@
-package cz.cvut.fit.gatling.protocol
-
-import io.gatling.core.protocol.ProtocolComponents
-import io.gatling.core.session.Session
-
-case class GremlinComponents(gremlinProtocol: GremlinProtocol) extends ProtocolComponents {
-  override def onStart: Option[(Session) => Session] = None
-  override def onExit: Option[(Session) => Unit] = None
-}
