--- src/main/scala/cz.cvut.fit/gatling/protocol/Utility.java
@@ -1,11 +1,0 @@
-package cz.cvut.fit.gatling.protocol;
-
-/**
- * Created on 6/21/2017.
- *
- * @author Marek.Cervak
- */
-public class Utility {
-
-
-}
