+++ src/main/scala/cz/cvut/fit/cervamar/gatling/check/ResultCheckSupport.scala
@@ -1,0 +1,10 @@
+package cz.cvut.fit.cervamar.gatling.check
+
+/**
+  * Created on 12/26/2017.
+  *
+  * @author Marek.Cervak
+  */
+trait ResultCheckSupport {
+    def simpleCheck = SimpleResultCheck
+}
