--- src/test/resources/neo4j-remote.properties
@@ -1,2 +1,0 @@
-gremlin.graph=com.steelbridgelabs.oss.neo4j.structure.Neo4JGraphFactory
-cz.cvut.fit.gremlin.sources.ElementIdProvider
