+++ src/test/scala/cz/cvut/fit/cervamar/gatling/simulations/HttpSimulation1.scala
@@ -1,0 +1,23 @@
+package cz.cvut.fit.cervamar.gatling.simulations
+
+import io.gatling.core.Predef._
+import io.gatling.http.Predef._
+import io.gatling.http.request.builder.HttpRequestBuilder.toActionBuilder
+
+/**
+  * Example Gatling load test that sends two HTTP requests to the same URL.
+  */
+class HttpSimulation1 extends Simulation {
+
+  val theHttpProtocolBuilder = http
+    .baseURL("http://computer-database.gatling.io")
+
+  val theScenarioBuilder = scenario("Scenario1")
+    .exec(
+      http("myRequest1").get("/")
+    )
+
+  setUp(
+    theScenarioBuilder.inject(atOnceUsers(1))
+  ).protocols(theHttpProtocolBuilder)
+}
